+++ mantis/cli/main.py
@@ -5,10 +5,11 @@ from mantis.cli.deskew import deskew
 from mantis.cli.estimate_affine import estimate_affine
 from mantis.cli.estimate_bleaching import estimate_bleaching
 from mantis.cli.estimate_deskew import estimate_deskew
+from mantis.cli.estimate_stabilization import estimate_stabilization
 from mantis.cli.optimize_affine import optimize_affine
 from mantis.cli.run_acquisition import run_acquisition
+from mantis.cli.stabilize import apply_stabilization
 from mantis.cli.update_scale_metadata import update_scale_metadata
-from mantis.cli.stabilization import estimate_stabilization_affine_list, stabilize_timelapse
 
 CONTEXT = {"help_option_names": ["-h", "--help"]}
 
@@ -32,5 +33,5 @@ cli.add_command(estimate_affine)
 cli.add_command(optimize_affine)
 cli.add_command(apply_affine)
 cli.add_command(update_scale_metadata)
-cli.add_command(estimate_stabilization_affine_list)
-cli.add_command(stabilize_timelapse)
+cli.add_command(estimate_stabilization)
+cli.add_command(apply_stabilization)
